requirements.txt
pygame
python-daemon
tk
asterisk-ami
